Better ROOTFile show()
