+++ src/UnROOT.jl
@@ -3,11 +3,12 @@ module UnROOT
 export ROOTFile, LazyBranch, LazyTree
 
 import Base: keys, get, getindex, show, length, iterate, position, ntoh, lock, unlock
+import AbstractTrees: children, printnode, print_tree
 using Base.Threads: SpinLock
 ntoh(b::Bool) = b
 
 using CodecZlib, CodecLz4, CodecXz, CodecZstd, StaticArrays
-using Mixers, Parameters, Memoization, LRUCache 
+using Mixers, Parameters, Memoization, LRUCache
 import Tables, TypedTables, PrettyTables
 
 @static if VERSION < v"1.1"
@@ -27,6 +28,7 @@ include("bootstrap.jl")
 include("root.jl")
 include("iteration.jl")
 include("custom.jl")
+include("displays.jl")
 
 
 end # module
